requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
celery~=5.3.6
redis~=5.0.1
pydantic>=2.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-celery>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
